requirements.txt

python-telegram-bot==21.2
Flask
requests
nest_asyncio
